finalize en-/decoding
