+++ csv/generic/src/main/scala/fs2/data/csv/generic/SeqShapedRowEncoder.scala
@@ -16,7 +16,7 @@
 package fs2.data.csv.generic
 
 import cats.data.NonEmptyList
-import fs2.data.csv.{CellEncoder, RowEncoder}
+import fs2.data.csv.{CellEncoder, Row, RowEncoder}
 import shapeless._
 
 trait SeqShapedRowEncoder[Repr] extends RowEncoder[Repr]
@@ -24,11 +24,11 @@ trait SeqShapedRowEncoder[Repr] extends RowEncoder[Repr]
 object SeqShapedRowEncoder {
 
   implicit def lastElemEncoder[Head](implicit Head: CellEncoder[Head]): SeqShapedRowEncoder[Head :: HNil] =
-    (last: Head :: HNil) => NonEmptyList.one(Head(last.head))
+    (last: Head :: HNil) => Row(NonEmptyList.one(Head(last.head)))
 
   implicit def hconsEncoder[Head, Tail <: HList](implicit
       Head: CellEncoder[Head],
       Tail: Lazy[SeqShapedRowEncoder[Tail]]): SeqShapedRowEncoder[Head :: Tail] =
-    (fields: Head :: Tail) => NonEmptyList(Head(fields.head), Tail.value(fields.tail).toList)
+    (fields: Head :: Tail) => Row(NonEmptyList(Head(fields.head), Tail.value(fields.tail).values.toList))
 
 }
