+++ csv/generic/src/main/scala/fs2/data/csv/generic/hlist.scala
@@ -18,21 +18,14 @@ package data
 package csv
 package generic
 
-import cats.data.NonEmptyList
-
 import shapeless._
 
 object hlist {
 
   final implicit def hlistDecoder[T <: HList](implicit cc: Lazy[SeqShapedRowDecoder[T]]): DerivedRowDecoder[T] =
-    new DerivedRowDecoder[T] {
-      def apply(row: Row): DecoderResult[T] =
-        cc.value(row)
-    }
+    (row: Row) => cc.value(row)
 
   final implicit def hlistEncoder[T <: HList](implicit cc: Lazy[SeqShapedRowEncoder[T]]): DerivedRowEncoder[T] =
-    new DerivedRowEncoder[T] {
-      override def apply(elem: T): NonEmptyList[String] = cc.value(elem)
-    }
+    (elem: T) => cc.value(elem)
 
 }
