+++ csv/shared/src/main/scala/fs2/data/csv/HasHeaders.scala
@@ -1,6 +1,6 @@
 package fs2.data.csv
 
-sealed trait HasHeaders[H[a] <: Option[a], Header] extends (RowF[H, Header] => CsvRow[Header])
+sealed trait HasHeaders[H[+a] <: Option[a], Header] extends (RowF[H, Header] => CsvRow[Header])
 
 object HasHeaders {
   implicit def hasHeaders[Header]: HasHeaders[Some, Header] = new HasHeaders[Some, Header] {
