+++ csv/shared/src/main/scala/fs2/data/csv/Row.scala
@@ -1,0 +1,8 @@
+package fs2.data.csv
+
+import cats.data.NonEmptyList
+
+object Row {
+  def apply(values: NonEmptyList[String]): Row = new Row(values, None)
+  def unapply(arg: Row): Some[NonEmptyList[String]] = Some(arg.values)
+}
