+++ csv/shared/src/main/scala/fs2/data/csv/RowDecoderF.scala
@@ -1,0 +1,124 @@
+package fs2.data.csv
+
+import cats._
+import cats.implicits._
+
+import scala.annotation.tailrec
+
+/** Describes how a row can be decoded to the given type.
+  *
+  * `RowDecoderF` provides convenient methods such as `map`, `emap`, or `flatMap`
+  * to build new decoders out of more basic one.
+  *
+  * Actually, `RowDecoderF` has a [[https://typelevel.org/cats/api/cats/MonadError.html cats `MonadError`]]
+  * instance. To get the full power of it, import `cats.implicits._`.
+  */
+trait RowDecoderF[H[+a] <: Option[a], T, Header] {
+  def apply(row: RowF[H, Header]): DecoderResult[T]
+
+  /**
+    * Map the parsed value.
+    * @param f the mapping function
+    * @tparam T2 the result type
+    * @return a row decoder reading the mapped type
+    */
+  def map[T2](f: T => T2): RowDecoderF[H, T2, Header] =
+    row => apply(row).map(f)
+
+  /**
+    * Map the parsed value to a new decoder, which in turn will be applied to
+    * the parsed value.
+    * @param f the mapping function
+    * @tparam T2 the result type
+    * @return a row decoder reading the mapped type
+    */
+  def flatMap[T2](f: T => RowDecoderF[H, T2, Header]): RowDecoderF[H, T2, Header] =
+    row => apply(row).flatMap(f(_)(row))
+
+  /**
+    * Map the parsed value, potentially failing.
+    * @param f the mapping function
+    * @tparam T2 the result type
+    * @return a row decoder reading the mapped type
+    */
+  def emap[T2](f: T => DecoderResult[T2]): RowDecoderF[H, T2, Header] =
+    row => apply(row).flatMap(f)
+
+  /**
+    * Fail-over. If this decoder fails, try the supplied other decoder.
+    * @param cd the fail-over decoder
+    * @tparam TT the return type
+    * @return a decoder combining this and the other decoder
+    */
+  def or[TT >: T](cd: => RowDecoderF[H, TT, Header]): RowDecoderF[H, TT, Header] =
+    row =>
+      apply(row) match {
+        case Left(_)      => cd(row)
+        case r @ Right(_) => r.leftCast[DecoderError]
+      }
+
+  /**
+    * Similar to [[or]], but return the result as an Either signaling which row decoder succeeded. Allows for parsing
+    * an unrelated type in case of failure.
+    * @param cd the alternative decoder
+    * @tparam B the type the alternative decoder returns
+    * @return a decoder combining both decoders
+    */
+  def either[B](cd: RowDecoderF[H, B, Header]): RowDecoderF[H, Either[T, B], Header] =
+    row =>
+      apply(row) match {
+        case Left(_) =>
+          cd(row) match {
+            case l @ Left(_)  => l.rightCast[Either[T, B]]
+            case r @ Right(_) => r.leftCast[T].asRight
+          }
+        case Right(value) => Right(Left(value))
+      }
+}
+
+object RowDecoderF extends ExportedRowDecoderFs {
+
+  implicit def RowDecoderFInstances[H[+a] <: Option[a], Header]
+      : MonadError[RowDecoderF[H, *, Header], DecoderError] with SemigroupK[RowDecoderF[H, *, Header]] =
+    new MonadError[RowDecoderF[H, *, Header], DecoderError] with SemigroupK[RowDecoderF[H, *, Header]] {
+      override def map[A, B](fa: RowDecoderF[H, A, Header])(f: A => B): RowDecoderF[H, B, Header] =
+        fa.map(f)
+
+      def flatMap[A, B](fa: RowDecoderF[H, A, Header])(f: A => RowDecoderF[H, B, Header]): RowDecoderF[H, B, Header] =
+        fa.flatMap(f)
+
+      def handleErrorWith[A](fa: RowDecoderF[H, A, Header])(
+          f: DecoderError => RowDecoderF[H, A, Header]): RowDecoderF[H, A, Header] =
+        row => fa(row).leftFlatMap(f(_)(row))
+
+      def pure[A](x: A): RowDecoderF[H, A, Header] =
+        _ => Right(x)
+
+      def raiseError[A](e: DecoderError): RowDecoderF[H, A, Header] =
+        _ => Left(e)
+
+      def tailRecM[A, B](a: A)(f: A => RowDecoderF[H, Either[A, B], Header]): RowDecoderF[H, B, Header] = {
+        @tailrec
+        def step(row: RowF[H, Header], a: A): DecoderResult[B] =
+          f(a)(row) match {
+            case left @ Left(_)          => left.rightCast[B]
+            case Right(Left(a))          => step(row, a)
+            case Right(right @ Right(_)) => right.leftCast[DecoderError]
+          }
+        row => step(row, a)
+      }
+
+      def combineK[A](x: RowDecoderF[H, A, Header], y: RowDecoderF[H, A, Header]): RowDecoderF[H, A, Header] = x or y
+    }
+
+}
+
+trait ExportedRowDecoderFs {
+  /*implicit def exportedRowDecoderFs[H[+a] <: Option[a], A, Header](implicit
+      exported: Exported[RowDecoderF[H, A, Header]]): RowDecoderF[H, A, Header] = exported.instance*/
+
+  implicit def exportedRowDecoders[A](implicit exported: Exported[RowDecoder[A]]): RowDecoder[A] = exported.instance
+
+  implicit def exportedCsvRowDecoders[A](implicit
+      exported: Exported[CsvRowDecoder[A, String]]): CsvRowDecoder[A, String] = exported.instance
+}
