+++ csv/shared/src/main/scala/fs2/data/csv/RowEncoder.scala
@@ -15,36 +15,13 @@
  */
 package fs2.data.csv
 
-import cats._
 import cats.data.NonEmptyList
 
-import scala.annotation.implicitNotFound
-
-/** Describes how a row can be encoded from a value of the given type.
-  */
-@implicitNotFound(
-  "No implicit RowEncoder found for type ${T}.\nYou can define one using RowEncoder.instance, by calling contramap on another RowEncoder or by using generic derivation for product types like case classes.\nFor that, add the fs2-data-csv-generic module to your dependencies and use either full-automatic derivation:\nimport fs2.data.csv.generic.auto._\nor the recommended semi-automatic derivation:\nimport fs2.data.csv.generic.semiauto._\nimplicit val rowEncoder: RowEncoder[${T}] = deriveRowEncoder\nMake sure to have instances of CellEncoder for every member type in scope.\n")
-trait RowEncoder[T] {
-  def apply(elem: T): NonEmptyList[String]
-
-  def contramap[B](f: B => T): RowEncoder[B] = elem => apply(f(elem))
-}
-
-object RowEncoder extends ExportedRowEncoders {
-
-  implicit object RowEncoderContravariant extends Contravariant[RowEncoder] {
-    override def contramap[A, B](fa: RowEncoder[A])(f: B => A): RowEncoder[B] =
-      fa.contramap(f)
-  }
+object RowEncoder {
 
   @inline
   def apply[T: RowEncoder]: RowEncoder[T] = implicitly[RowEncoder[T]]
 
   @inline
-  def instance[T](f: T => NonEmptyList[String]): RowEncoder[T] = f(_)
-}
-
-trait ExportedRowEncoders {
-  implicit def exportedRowEncoders[A](implicit exported: Exported[RowEncoder[A]]): RowEncoder[A] =
-    exported.instance
+  def instance[T](f: T => NonEmptyList[String]): RowEncoder[T] = (t: T) => Row(f(t))
 }
