+++ csv/shared/src/main/scala/fs2/data/csv/package.scala
@@ -26,107 +26,25 @@ import scala.annotation.{implicitNotFound, tailrec}
 
 package object csv {
 
-  trait RowDecoderF[H[a] <: Option[a], T, Header] {
-    def apply(row: RowF[H, Header]): DecoderResult[T]
+  /**
+    * Higher kinded version of [[scala.None]]. Ignores the type param.
+    */
+  type NoneF[+A] = None.type
 
-    /**
-      * Map the parsed value.
-      * @param f the mapping function
-      * @tparam T2 the result type
-      * @return a row decoder reading the mapped type
-      */
-    def map[T2](f: T => T2): RowDecoderF[H, T2, Header] =
-      row => apply(row).map(f)
+  /**
+    * A CSV row without headers.
+    */
+  type Row = RowF[NoneF, Nothing]
 
-    /**
-      * Map the parsed value to a new decoder, which in turn will be applied to
-      * the parsed value.
-      * @param f the mapping function
-      * @tparam T2 the result type
-      * @return a row decoder reading the mapped type
-      */
-    def flatMap[T2](f: T => RowDecoderF[H, T2, Header]): RowDecoderF[H, T2, Header] =
-      row => apply(row).flatMap(f(_)(row))
+  /** A CSV row with headers, that can be used to access the cell values.
+    *
+    * '''Note:''' the following invariant holds when using this class: `values` and `headers` have the same size.
+    */
+  type CsvRow[Header] = RowF[Some, Header]
 
-    /**
-      * Map the parsed value, potentially failing.
-      * @param f the mapping function
-      * @tparam T2 the result type
-      * @return a row decoder reading the mapped type
-      */
-    def emap[T2](f: T => DecoderResult[T2]): RowDecoderF[H, T2, Header] =
-      row => apply(row).flatMap(f)
+  type HeaderResult[T] = Either[HeaderError, NonEmptyList[T]]
 
-    /**
-      * Fail-over. If this decoder fails, try the supplied other decoder.
-      * @param cd the fail-over decoder
-      * @tparam TT the return type
-      * @return a decoder combining this and the other decoder
-      */
-    def or[TT >: T](cd: => RowDecoderF[H, TT, Header]): RowDecoderF[H, TT, Header] =
-      row =>
-        apply(row) match {
-          case Left(_)      => cd(row)
-          case r @ Right(_) => r.leftCast[DecoderError]
-        }
-
-    /**
-      * Similar to [[or]], but return the result as an Either signaling which row decoder succeeded. Allows for parsing
-      * an unrelated type in case of failure.
-      * @param cd the alternative decoder
-      * @tparam B the type the alternative decoder returns
-      * @return a decoder combining both decoders
-      */
-    def either[B](cd: RowDecoderF[H, B, Header]): RowDecoderF[H, Either[T, B], Header] =
-      row =>
-        apply(row) match {
-          case Left(_) =>
-            cd(row) match {
-              case l @ Left(_)  => l.rightCast[Either[T, B]]
-              case r @ Right(_) => r.leftCast[T].asRight
-            }
-          case Right(value) => Right(Left(value))
-        }
-  }
-
-  object RowDecoderF extends ExportedCsvRowDecoders with ExportedRowDecoders {
-
-    implicit def RowDecoderFInstances[H[a] <: Option[a], Header]
-        : MonadError[RowDecoderF[H, *, Header], DecoderError] with SemigroupK[RowDecoderF[H, *, Header]] =
-      new MonadError[RowDecoderF[H, *, Header], DecoderError] with SemigroupK[RowDecoderF[H, *, Header]] {
-        override def map[A, B](fa: RowDecoderF[H, A, Header])(f: A => B): RowDecoderF[H, B, Header] =
-          fa.map(f)
-
-        def flatMap[A, B](fa: RowDecoderF[H, A, Header])(f: A => RowDecoderF[H, B, Header]): RowDecoderF[H, B, Header] =
-          fa.flatMap(f)
-
-        def handleErrorWith[A](fa: RowDecoderF[H, A, Header])(
-            f: DecoderError => RowDecoderF[H, A, Header]): RowDecoderF[H, A, Header] =
-          row => fa(row).leftFlatMap(f(_)(row))
-
-        def pure[A](x: A): RowDecoderF[H, A, Header] =
-          _ => Right(x)
-
-        def raiseError[A](e: DecoderError): RowDecoderF[H, A, Header] =
-          _ => Left(e)
-
-        def tailRecM[A, B](a: A)(f: A => RowDecoderF[H, Either[A, B], Header]): RowDecoderF[H, B, Header] = {
-          @tailrec
-          def step(row: RowF[H, Header], a: A): DecoderResult[B] =
-            f(a)(row) match {
-              case left @ Left(_)          => left.rightCast[B]
-              case Right(Left(a))          => step(row, a)
-              case Right(right @ Right(_)) => right.leftCast[DecoderError]
-            }
-          row => step(row, a)
-        }
-
-        def combineK[A](x: RowDecoderF[H, A, Header], y: RowDecoderF[H, A, Header]): RowDecoderF[H, A, Header] = x or y
-      }
-
-  }
-
-  type NoneF[A] = None.type
+  type DecoderResult[T] = Either[DecoderError, T]
 
   /** Describes how a row can be decoded to the given type.
     *
@@ -140,6 +58,10 @@ package object csv {
     "No implicit RowDecoder found for type ${T}.\nYou can define one using RowDecoder.instance, by calling map on another RowDecoder or by using generic derivation for product types like case classes.\nFor that, add the fs2-data-csv-generic module to your dependencies and use either full-automatic derivation:\nimport fs2.data.csv.generic.auto._\nor the recommended semi-automatic derivation:\nimport fs2.data.csv.generic.semiauto._\nimplicit val rowDecoder: RowDecoder[${T}] = deriveRowDecoder\nMake sure to have instances of CellDecoder for every member type in scope.\n")
   type RowDecoder[T] = RowDecoderF[NoneF, T, Nothing]
 
+  /** Describes how a row can be encoded from a value of the given type.
+    */
+  @implicitNotFound(
+    "No implicit RowEncoder found for type ${T}.\nYou can define one using RowEncoder.instance, by calling contramap on another RowEncoder or by using generic derivation for product types like case classes.\nFor that, add the fs2-data-csv-generic module to your dependencies and use either full-automatic derivation:\nimport fs2.data.csv.generic.auto._\nor the recommended semi-automatic derivation:\nimport fs2.data.csv.generic.semiauto._\nimplicit val rowEncoder: RowEncoder[${T}] = deriveRowEncoder\nMake sure to have instances of CellEncoder for every member type in scope.\n")
   type RowEncoder[T] = RowEncoderF[NoneF, T, Nothing]
 
   /** Describes how a row can be decoded to the given type.
@@ -154,48 +76,11 @@ package object csv {
     "No implicit CsvRowDecoder found for type ${T}.\nYou can define one using CsvRowDecoder.instance, by calling map on another CsvRowDecoder or by using generic derivation for product types like case classes.\nFor that, add the fs2-data-csv-generic module to your dependencies and use either full-automatic derivation:\nimport fs2.data.csv.generic.auto._\nor the recommended semi-automatic derivation:\nimport fs2.data.csv.generic.semiauto._\nimplicit val csvRowDecoder: CsvRowDecoder[${T}] = deriveCsvRowDecoder\nMake sure to have instances of CellDecoder for every member type in scope.\n")
   type CsvRowDecoder[T, Header] = RowDecoderF[Some, T, Header]
 
+  /** Describes how a row can be encoded from a value of the given type.
+    */
+  @implicitNotFound(
+    "No implicit CsvRowEncoderF[H,  found for type ${T}.\nYou can define one using CsvRowEncoderF[H, .instance, by calling contramap on another CsvRowEncoderF[H,  or by using generic derivation for product types like case classes.\nFor that, add the fs2-data-csv-generic module to your dependencies and use either full-automatic derivation:\nimport fs2.data.csv.generic.auto._\nor the recommended semi-automatic derivation:\nimport fs2.data.csv.generic.semiauto._\nimplicit val csvRowEncoder: CsvRowEncoderF[H, [${T}] = deriveCsvRowEncoderF[H, \nMake sure to have instances of CellEncoder for every member type in scope.\n")
   type CsvRowEncoder[T, Header] = RowEncoderF[Some, T, Header]
-
-  type Row = RowF[NoneF, Nothing]
-  object Row {
-    def apply(values: NonEmptyList[String]): Row = new Row(values, None)
-    def unapply(arg: Row): Some[NonEmptyList[String]] = Some(arg.values)
-  }
-
-  type CsvRow[Header] = RowF[Some, Header]
-
-  object CsvRow {
-
-    /** Constructs a [[CsvRow]] and checks that the size of values and headers match. */
-    def apply[Header](values: NonEmptyList[String],
-                      headers: NonEmptyList[Header]): Either[CsvException, CsvRow[Header]] =
-      if (values.length =!= headers.length)
-        Left(
-          new CsvException(
-            s"Headers have size ${headers.length} but row has size ${values.length}. Both numbers must match!"))
-      else
-        Right(new CsvRow(values, Some(headers)))
-
-    def unsafe[Header](values: NonEmptyList[String], headers: NonEmptyList[Header]): CsvRow[Header] =
-      apply(values, headers).fold(throw _, identity)
-
-    def fromListHeaders[Header](l: List[(Header, String)]): Option[CsvRow[Header]] = {
-      val (hs, vs) = l.unzip
-      (NonEmptyList.fromList(vs), NonEmptyList.fromList(hs)).mapN((v, h) => new CsvRow(v, Some(h)))
-    }
-
-    def fromNelHeaders[Header](nel: NonEmptyList[(Header, String)]): CsvRow[Header] = {
-      val (hs, vs) = nel.toList.unzip
-      new CsvRow(NonEmptyList.fromListUnsafe(vs), Some(NonEmptyList.fromListUnsafe(hs)))
-    }
-
-    def unapply[Header](arg: CsvRow[Header]): Some[(NonEmptyList[String], NonEmptyList[Header])] = Some(
-      (arg.values, arg.headers.get))
-  }
-
-  type HeaderResult[T] = Either[HeaderError, NonEmptyList[T]]
-
-  type DecoderResult[T] = Either[DecoderError, T]
 
   sealed trait QuoteHandling
   object QuoteHandling {
